--- Project/GAMZ/src/main/java/com/app/App.java
@@ -1,14 +1,0 @@
-package com.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class App {
-
-    public static void main(String[] args) {
-
-        SpringApplication.run(App.class, args);
-    }
-}
