--- Project/GAMZ/src/main/java/com/app/project/Admin.java
@@ -1,9 +1,0 @@
-package com.app.project;
-
-import javax.persistence.Entity;
-import javax.persistence.Table;
-
-@Table
-@Entity
-public class Admin extends User {
-}
