--- Project/GAMZ/src/main/java/com/app/project/Faculty.java
@@ -1,17 +1,0 @@
-package com.app.project;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Table
-@Entity
-public class Faculty {
-    @Id
-    @GeneratedValue()
-    private int id;
-    private String name;
-    //dose this mean 2 indexes then
-
-    @OneToMany
-    private List<Professor> allProfessors;
-}
