--- Project/GAMZ/src/main/java/com/app/project/Field.java
@@ -1,12 +1,0 @@
-package com.app.project;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Table
-@Entity
-public class Field {
-    @Id
-    private String field;
-}
