--- Project/GAMZ/src/main/java/com/app/project/ProfessorRecomendation.java
@@ -1,8 +1,0 @@
-package com.app.project;
-
-enum ProfessorRecomendation {
-    NOT_RECOMMENDED,
-    RECOMEND_NO_SUPERVISION,
-    RECOMEND_NO_FUND,
-    RECOMEND_FUND
-}
