--- Project/GAMZ/src/main/java/com/app/project/Project.java
@@ -1,25 +1,0 @@
-package com.app.project;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import java.util.List;
-@Table
-@Entity
-public class Project {
-    @Id
-    private int id;
-    @OneToMany
-    private List<Applicant> approvedStudents;
-    private String projectDescription;
-    @OneToMany
-    private List<Field> projectsFeildsOfApplication;
-    @OneToMany
-    private List<Field> projectFeildsRequired;
-
-    private boolean activeProject;
-    private boolean approval;
-
-
-}
