--- Project/GAMZ/src/main/java/com/app/project/Skills.java
@@ -1,15 +1,0 @@
-package com.app.project;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Table
-@Entity
-public class Skills {
-
-    @Id
-    @GeneratedValue
-    private int id;
-}
