--- Project/GAMZ/src/main/java/com/app/project/User.java
@@ -1,45 +1,0 @@
-package com.app.project;
-
-import javax.persistence.*;
-
-@Table
-@Entity
-public class User {
-    @Id
-    @GeneratedValue()
-    private int id;
-   //@Index()
-    private String name;
-    private String role;
-    private String number;
-   // @Index()
-    private String email;
-   // @Index()
-    private String url1;
-   // @Index()
-    private String url2;
-
-    public User(){}
-    public User(String name){
-        this.name = name;
-    }
-
-    public User(String name, String role){
-        this.name = name;
-        this.role = role;
-    }
-
-    public Integer getId(){  return id; }
-
-    public String getName(){  return name; }
-
-    public void setName(String name){
-        this.name = name;
-    }
-
-    public String getRole(){
-        return role;
-    }
-
-
-}
