+++ src/main/java/Feild.java
@@ -1,7 +1,10 @@
 import javax.persistence.Entity;
+import javax.persistence.Id;
 import javax.persistence.Table;
 
 @Table
 @Entity
 public class Feild {
+    @Id
+    private String feild;
 }
