+++ src/main/java/User.java
@@ -6,14 +6,37 @@ public class User {
     @Id
     @GeneratedValue()
     private int id;
-    @Index()
+    //@Index()
     private String name;
-
+    private String role;
     private String number;
-    @Index()
+    //@Index()
     private String email;
-    @Index()
+    //@Index()
     private String url1;
-    @Index()
+    //@Index()
     private String url2;
+
+
+    public User(){}
+    public User(String name){
+        this.name = name;
+    }
+
+    public User(String name, String role){
+        this.name = name;
+        this.role = role;
+    }
+
+    public Integer getId(){  return id; }
+
+    public String getName(){  return name; }
+
+    public void setName(String name){
+        this.name = name;
+    }
+
+    public String getRole(){
+        return role;
+    }
 }
