+++ src/main/resources/templates/student.html
